Fix Psionic Invisibility
