+++ Content.Server/Abilities/Psionics/Abilities/MindSwapPowerSystem.cs
@@ -8,7 +8,7 @@ using Content.Shared.Damage;
 using Content.Server.Mind;
 using Content.Shared.Mobs.Systems;
 using Content.Server.Popups;
-using Content.Server.Psionics;
+using Content.Shared.Psionics;
 using Content.Server.GameTicking;
 using Content.Shared.Mind;
 using Content.Shared.Actions.Events;
