--- Content.Server/Psionics/Invisibility/PsionicallyInvisibleComponent.cs
@@ -1,6 +1,0 @@
-namespace Content.Server.Psionics
-{
-    [RegisterComponent]
-    public sealed partial class PsionicallyInvisibleComponent : Component
-    {}
-}
