+++ Content.Shared/Psionics/Abilities/PsionicInvisibility/PsionicallyInvisibleComponent.cs
@@ -1,0 +1,4 @@
+namespace Content.Shared.Psionics;
+
+[RegisterComponent]
+public sealed partial class PsionicallyInvisibleComponent : Component { }
